supplychain/wsgi.py
"""
WSGI config for supplychain project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/5.1/howto/deployment/wsgi/
"""

import os
from pathlib import Path
from dotenv import load_dotenv

# Load .env for gunicorn/uWSGI etc.
load_dotenv(Path(__file__).resolve().parent.parent / ".env")

import inspect

if not hasattr(inspect, 'getargspec'):
    inspect.getargspec = inspect.getfullargspec

from django.core.wsgi import get_wsgi_application

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'supplychain.settings')

application = get_wsgi_application()


