requirements.txt
python-dotenv==1.0.1
phidata==2.7.10
openai>=1.0.0
tavily-python==0.5.1
Django==5.1.7
gunicorn==20.1.0
geopy==2.4.1
sqlalchemy==2.0.40
yfinance==0.2.55
